+++ packages/site/src/utils/metamask.ts
@@ -1,3 +1,8 @@
+import * as ethers from "ethers";
+import { Wallet__factory } from "@nocturne-xyz/contracts";
+
+const WALLET_ADDRESS = "0x9e4Fa3251ee437379398e486E9A27A997d90ce51";
+
 /**
  * Detect if the wallet injecting the ethereum object is Flask.
  *
@@ -17,4 +22,11 @@ export const isFlask = async () => {
   } catch {
     return false;
   }
+};
+
+export const connectWalletContract = async () => {
+  const provider = new ethers.providers.Web3Provider(window.ethereum as any);
+  await provider.send("eth_requestAccounts", []);
+  const signer = provider.getSigner();
+  return Wallet__factory.connect(WALLET_ADDRESS, signer);
 };