+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Unreleased
 
+- Depend on nocturne fork of circomlibjs
+- Fix `decodeAsset` not padding out address with zeros
+- Change `NocturneContext` constructor to take wallet address and provider instead of `Wallet` object
+- Use `randombytes` instead of node `crypto`
+- Use patched fork of `circomlibjs`
+- Avoid usage of `Buffer` in `sdk`.
 - Merge `joinSplitTx.encodedAssetAddr` and `joinSplitTx.encodedAssetId` into one field `joinSplitTx.encodedAsset`
 - Add utility function to estimate `executionGasLimit` and `maxNumRefund`
 - Change note commitment to use encoded `asset` and `ID`